+++ src/listeners.js
@@ -10,6 +10,7 @@ import {
   deleteSongFromPlaylist,
 } from "./helpers/index.js";
 
+import { presentationHandler } from "./admin/alavirien.js";
 import { roleAdd, roleRemove } from "./admin/role.js";
 
 // jsons imports
@@ -165,6 +166,13 @@ export const onReactionAdd = async (messageReaction, user) => {
   if (currentServer.octagonalSign === messageReaction.emoji.name) {
     octagonalLog(messageReaction, user);
     return;
+  }
+
+
+  if (messageReaction.message.channel.id === currentServer.presentationChannelId && currentServer.presentationReactId === messageReaction.emoji.name) {
+    console.log("detected")
+    presentationHandler(currentServer, messageReaction, user);
+    return; //no command in presentation channel
   }
 
   onRemoveSpotifyReaction(messageReaction, currentServer);